Moved motion_interface to intera_motion_interface
