+++ intera_interface/src/intera_motion_interface/interaction_options.py
