+++ intera_interface/src/intera_motion_interface/motion_trajectory.py
@@ -32,12 +32,12 @@ from intera_core_msgs.msg import (
 from sensor_msgs.msg import (
     JointState
 )
-from motion_interface.motion_controller_action_client import (
+from motion_controller_action_client import (
     MotionControllerActionClient
 )
-from motion_interface.motion_waypoint import MotionWaypoint
-from motion_interface.motion_waypoint_options import MotionWaypointOptions
-from motion_interface.utility_functions import ensure_path_to_file_exists
+from motion_waypoint import MotionWaypoint
+from motion_waypoint_options import MotionWaypointOptions
+from utility_functions import ensure_path_to_file_exists
 from copy import deepcopy
 from rospy_message_converter import message_converter
