+++ intera_interface/src/intera_motion_interface/motion_waypoint.py
@@ -29,7 +29,7 @@ from copy import deepcopy
 from utility_functions import ensure_path_to_file_exists
 from rospy_message_converter import message_converter
 import yaml
-from motion_interface.motion_waypoint_options import MotionWaypointOptions
+from motion_waypoint_options import MotionWaypointOptions
 from intera_interface import Limb
 
 class MotionWaypoint(object):
